+++ src/classes/managers/authentication/errors/application-auth.error.ts
@@ -1,7 +1,9 @@
 import { cTraderXError } from '../../../models/ctrader-x-error.model';
 
 export class ApplicationAuthenticationError extends cTraderXError {
-    constructor(message: string) {
-        super(`Application authentication error: ${message}`);
+    constructor(error: unknown) {
+        super(
+            `Application authentication error: ${cTraderXError.getMessageError(error)}`,
+        );
     }
 }