+++ src/classes/managers/symbols/errors/get-symbols-list.error.ts
@@ -1,7 +1,9 @@
 import { cTraderXError } from '../../../models/ctrader-x-error.model';
 
 export class GetSymbolsListError extends cTraderXError {
-    constructor(message: string) {
-        super(`Get symbols list error: ${message}`);
+    constructor(error: unknown) {
+        super(
+            `Get symbols list error: ${cTraderXError.getMessageError(error)}`,
+        );
     }
 }