+++ src/classes/managers/symbols/errors/get-trend-bars.error.ts
@@ -1,0 +1,7 @@
+import { cTraderXError } from '../../../models/ctrader-x-error.model';
+
+export class GetTrendBarsError extends cTraderXError {
+    constructor(readonly error: unknown) {
+        super(`Get trend bars error: ${cTraderXError.getMessageError(error)}`);
+    }
+}