+++ src/classes/managers/symbols/symbols.manager.ts
@@ -5,11 +5,20 @@ import { ProtoOASymbolsListReq } from '../../../models/proto/messages/symbols/Pr
 import { ICredentials } from '../models/credentials.model';
 import { ProtoOASymbolsListRes } from '../../../models/proto/messages/symbols/ProtoOASymbolsListRes';
 import { GetSymbolsListError } from './errors/get-symbols-list.error';
-import { cTraderXError } from '../../models/ctrader-x-error.model';
+import { ProtoOAGetTrendbarsRes } from '../../../models/proto/messages/symbols/ProtoOAGetTrendbarsRes';
+import { ProtoOAGetTrendbarsReq } from '../../../models/proto/messages/symbols/ProtoOAGetTrendbarsReq';
+import { GetTrendBarsError } from './errors/get-trend-bars.error';
+import { ProtoOASymbolByIdRes } from '../../../models/proto/messages/symbols/ProtoOASymbolByIdRes';
+import { ProtoOASymbolByIdReq } from '../../../models/proto/messages/symbols/ProtoOASymbolByIdReq';
+import { GetSymbolsDetailsError } from './errors/get-symbols-details.error';
+import { ProtoOASymbol } from '../../../models/proto/models/ProtoOASymbol';
+import { ProtoOALightSymbol } from '../../../models/proto/models/ProtoOALightSymbol';
 
 export interface IGetSymbolsListOptions {
     includeArchivedSymbols?: boolean;
 }
+
+export type GetSymbolsListResult = (ProtoOASymbol & ProtoOALightSymbol)[];
 
 export class SymbolsManager extends BaseManager {
     constructor(
@@ -20,7 +29,9 @@ export class SymbolsManager extends BaseManager {
         super();
     }
 
-    async getSymbolsList(opts?: IGetSymbolsListOptions): Promise<ProtoOASymbolsListRes> {
+    async getSymbolsList(
+        opts?: IGetSymbolsListOptions,
+    ): Promise<GetSymbolsListResult> {
         this.logCallAttempt(this.getSymbolsList);
 
         const payload: ProtoOASymbolsListReq = {
@@ -35,12 +46,75 @@ export class SymbolsManager extends BaseManager {
                 payload,
             )) as ProtoOASymbolsListRes;
         } catch (e) {
-            const message = cTraderXError.getMessageError(e);
-            this.logCallAttemptFailure(this.getSymbolsList, message);
-            throw new GetSymbolsListError(message);
+            throw this.handleCTraderCallError(
+                e,
+                this.getSymbolsList,
+                new GetSymbolsListError(e),
+            );
         }
 
         this.logCallAttemptSuccess(this.getSymbolsList);
+
+        const fullSymbols = await this.getSymbolsDetails(
+            result.symbol.map(({ symbolId }) => symbolId),
+        );
+
+        return fullSymbols.symbol.map((symbol) => ({
+            ...symbol,
+            ...result.symbol.find((s) => s.symbolId === symbol.symbolId),
+        }));
+    }
+
+    async getSymbolsDetails(
+        symbolsIds: number[],
+    ): Promise<ProtoOASymbolByIdRes> {
+        this.logCallAttempt(this.getSymbolsDetails);
+
+        const payload: ProtoOASymbolByIdReq = {
+            symbolId: symbolsIds,
+        };
+
+        let result: ProtoOASymbolByIdRes;
+        try {
+            result = (await this.connection.sendCommand(
+                ProtoOASymbolByIdReq.name,
+                payload,
+            )) as ProtoOASymbolByIdRes;
+        } catch (e) {
+            throw this.handleCTraderCallError(
+                e,
+                this.getSymbolsDetails,
+                new GetSymbolsDetailsError(e),
+            );
+        }
+
+        this.logCallAttemptSuccess(this.getSymbolsDetails);
+        return result;
+    }
+
+    async getTrendBars(
+        opts: Omit<ProtoOAGetTrendbarsReq, 'ctidTraderAccountId'>,
+    ): Promise<ProtoOAGetTrendbarsRes> {
+        this.logCallAttempt(this.getTrendBars);
+        let result: ProtoOAGetTrendbarsRes;
+        try {
+            const payload: ProtoOAGetTrendbarsReq = {
+                ...opts,
+                ctidTraderAccountId: this.credentials.ctidTraderAccountId,
+            };
+            result = (await this.connection.sendCommand(
+                ProtoOAGetTrendbarsReq.name,
+                payload,
+            )) as ProtoOAGetTrendbarsRes;
+        } catch (e) {
+            throw this.handleCTraderCallError(
+                e,
+                this.getTrendBars,
+                new GetTrendBarsError(e),
+            );
+        }
+
+        this.logCallAttemptSuccess(this.getTrendBars);
         return result;
     }
 }