+++ src/models/proto/base-proto.ts
@@ -4,4 +4,10 @@ export class BaseProto {
      * https://help.ctrader.com/open-api/model-messages/#protooaordertype
      */
     payloadType?: number;
+
+    /**
+     * The unique identifier of the trader's account in cTrader platform.
+     * Used to match responses to trader's accounts.
+     */
+    ctidTraderAccountId?: number;
 }