+++ src/models/proto/messages/authentication/ProtoOAAccountAuthRes.ts
@@ -1,8 +1,3 @@
 import { BaseProto } from '../../base-proto';
 
-export class ProtoOAAccountAuthRes extends BaseProto {
-    /**
-     * The unique identifier of the trader's account in cTrader platform.
-     */
-    ctidTraderAccountId: number;
-}
+export class ProtoOAAccountAuthRes extends BaseProto {}