+++ src/models/proto/messages/symbols/ProtoOASymbolByIdReq.ts
@@ -1,0 +1,8 @@
+import { BaseProto } from '../../base-proto';
+
+export class ProtoOASymbolByIdReq extends BaseProto {
+    /**
+     * Unique identifier of the symbols in cTrader platform.
+     */
+    symbolId: number[];
+}