+++ src/models/proto/messages/symbols/ProtoOASymbolsListRes.ts
@@ -3,11 +3,6 @@ import { ProtoOALightSymbol } from '../../models/ProtoOALightSymbol';
 
 export class ProtoOASymbolsListRes extends BaseProto {
     /**
-     * Unique identifier of the trader's account. Used to match responses to trader's accounts.
-     */
-    ctidTraderAccountId: number;
-
-    /**
      * The list of symbols.
      */
     symbol: ProtoOALightSymbol[];