+++ src/models/proto/models/ProtoOATrendbar.ts
@@ -1,0 +1,11 @@
+import { ProtoOATrendbarPeriod } from './ProtoOATrendbarPeriod';
+
+export class ProtoOATrendbar {
+    volume: number;
+    period?: ProtoOATrendbarPeriod;
+    low?: number;
+    deltaOpen?: number;
+    deltaClose?: number;
+    deltaHigh?: number;
+    utcTimestampInMinutes?: number;
+}