+++ src/models/proto/models/ProtoOATrendbarPeriod.ts
@@ -1,0 +1,16 @@
+export enum ProtoOATrendbarPeriod {
+    M1 = 1,
+    M2 = 2,
+    M3 = 3,
+    M4 = 4,
+    M5 = 5,
+    M10 = 6,
+    M15 = 7,
+    M30 = 8,
+    H1 = 9,
+    H4 = 10,
+    H12 = 11,
+    D1 = 12,
+    W1 = 13,
+    MN1 = 1,
+}